Align cite groups & cite group delimiting to the spec.
Various things the tests were relying on:

- cite-group-delimiter (CGD) overriding even an explicitly provided
  year-suffix-delimiter (YSD), when the spec is pretty clear that YSD
  and after-collapse-delimiter (ACD) can be customised and default to LD
  (layout delimiter).
- Not using CGD when not specified, although it is meant to default to
  ", " rather than LD
- ACD being used between non-collapsed cites

This is best illustrated by a grouping that uses all of the delimiters,
in the new collapse_YearSuffixCollapseNoYearSuffixDelimiterNoFallback.txt.

citeproc-js renders the second line of that one as:

    Smith 2000a & c–e; 2001. Other 2002. More 2003

where LD = "; ", ACD = ". ", CGD = " & ", YSD = ","
Note that the year suffixes are not separated by YSD, YSD is not used at
all, and the roles of ACD and LD are kinda switched but not quite. Note
also that if you remove CGD from the <citation> element, you get

    Smith 2000a,c–e; 2001. Other 2002. More 2003

The new result is:

    Smith 2000a,c–e & 2001. Other 2002; More 2003

Here:
- YSD delimits single year-suffixes or ranges of year-suffixes only
  Crucially, these must have the same year, otherwise it can only be CGD.
- CGD delimit cites with a common name (Smith), if that delimiter
  is not already being handled as a YSD. CGD does not override YSD.
- ACD follows the group of collapsed Smith cites, and appears nowhere else
  (this accords with the only example of ACD in the spec)
- LD is used after regular cites that have not been collapsed at all

There are other fallback interactions shown by the other tests in the diff.